requirements.txt
django
djangorestframework
psycopg2
pgvector
redis
requests
streamlit
pandas
ollama